Fix attribution button color not applying on iOS/decoded incorrectly on Android

+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/mapping/AttributionMappings.kt
@@ -16,7 +16,7 @@ fun AttributionSettingsInterface.applyFromFLT(settings: FLTSettings.AttributionS
 }
 
 fun AttributionSettingsInterface.toFLT() = FLTSettings.AttributionSettings.Builder().let { settings ->
-  settings.setIconColor(iconColor.toLong())
+  settings.setIconColor(iconColor.toUInt().toLong())
   settings.setPosition(position.toOrnamentPosition())
   settings.setMarginLeft(marginLeft.toDouble())
   settings.setMarginTop(marginTop.toDouble())
